Move getLocalUser export fn, add copy to clipboard
