+++ src/utils/utils.js
@@ -1,0 +1,4 @@
+//Return user object form local storage
+export const getLocalUser = () => {
+    return { ...JSON.parse(localStorage.getItem("dinnerPokerUser")) }
+}
